Feature ETP-3981: Move LocationEditorModal to shared/ (F2)
Decouples generic PartnerAddressPicker (contract-ui) from the concrete contacts window. Adds a third resolution path in generate-frontend.js so customAddModal imports can fall back to windows/custom/shared/. Cleans up dead contacts/__tests__ glob in package.json.
