+++ tools/app-shell/package.json
@@ -8,7 +8,7 @@
     "dev:mock": "VITE_MOCK=true vite",
     "build": "vite build",
     "preview": "vite preview",
-    "test": "node --test src/lib/__tests__/*.test.js src/hooks/__tests__/*.test.js src/windows/__tests__/*.test.js src/components/ui/__tests__/*.test.js src/windows/custom/contacts/__tests__/*.test.js",
+    "test": "node --test src/lib/__tests__/*.test.js src/hooks/__tests__/*.test.js src/windows/__tests__/*.test.js src/components/ui/__tests__/*.test.js",
     "test:vitest": "vitest run",
     "test:vitest:coverage": "vitest run --coverage",
     "test:vitest:watch": "vitest",
